// src/models/teams/teamFinances.ts
export interface TeamFinances {
  ArenaCapacityL1: number;
  ArenaCapacityL2: number;
  ArenaCapacityL3: number;
  ArenaCapacityL4: number;
  ArenaCapacityLuxury: number;
  AttendanceL1: number;
  AttendanceL2: number;
  AttendanceL3: number;
  AttendanceL4: number;
  AttendanceLuxury: number;
  CurrentBankAccount: number;
  EstimatedRevenue: number;
  EstimatedSeasonExpense: number;
  ExpensePerDay: number;
  ExpenseThisSeason: number;
  FarmEstimatedSeasonExpense: number;
  HomeGP: number;
  LuxuryTaxeTotal: number;
  Name: string;
  Number: number;
  ProjectedBankAccount: number;
  SalaryCapPerDay: number;
  SalaryCapToDate: number;
  ScheduleGameInAYear: number;
  ScheduleHomeGameInAYear: number;
  SeasonTicketPCT: number;
  SpecialSalaryCapY1: number;
  SpecialSalaryCapY2: number;
  SpecialSalaryCapY3: number;
  SpecialSalaryCapY4: number;
  SpecialSalaryCapY5: number;
  SpecialSalaryCapY6: number;
  SpecialSalaryCapY7: number;
  SpecialSalaryCapY8: number;
  SpecialSalaryCapY9: number;
  SpecialSalaryCapY10: number;
  TeamPopularity: number;
  TicketPriceL1: number;
  TicketPriceL2: number;
  TicketPriceL3: number;
  TicketPriceL4: number;
  TicketPriceLuxury: number;
  TotalAttendance: number;
  TotalIncome: number;
  TotalPlayersSalaries: number;
  TotalPlayersSalariesAverage: number;
  TotalSalaryCap: number;
}